Added distance-to-top checker.
When the marked elements have reached their specified position, a new class name will be added to their class list.
This is especially useful when creating animations that appear on scroll.

To mark an element-> 
        To this elements class list add: 'scroll_marker' & 'scroll_offset_<integer_value'
The scroll offset value will change at which point on the screen the element will be marked. If it is not included, this value will be defaulted to 0px.